zadanie/app-web/requirements.txt
flask
kafka-python
flask_socketio